+++ database/migrations/2020_03_13_095750_add_fixed_quantity_to_posts_table.php
@@ -16,7 +16,7 @@ class AddFixedQuantityToPostsTable extends Migration
         Schema::table('posts', function (Blueprint $table) {
             //
 
-            $table->double('fixed_quantity');
+            $table->integer('fixed_quantity');
         });
     }
