+++ database/migrations/2020_03_13_095833_add_earnings_to_posts_table.php
@@ -16,7 +16,7 @@ class AddEarningsToPostsTable extends Migration
         Schema::table('posts', function (Blueprint $table) {
             //
 
-            $table->double('earnings');
+            $table->integer('earnings');
         });
     }
